Fixed issue with price calculation of awattar neu
The order how the prices are calculated has been wrong.

Currently it is calculated like (spot price + 1.5 Cent) + 3% + 20% and therefore a 3% surcharge on the 1.5Cent fee is added as well. This is wrong.

According to awattar price page, the 3% surcharge is only applied to the spot price but not to the additional fee.

I checked with my last invoice and after the fix, it matches the values.